Make changes to handling of input and output
Input can now come from --string, --password or a filename that is
the last argument on the commandline.  If none are supplied, the script
will assume it should try to read from stdin.

Output formatting can be overriden from the examples given to output
more script friendly format by setting --output to block or string.