Add eyaml file encryption and decryption
When using the --eyaml flag in conjunction with decrypt or encrypt
eyaml will now scan the file for sections wrapped in ENC[] and
decrypt or encrypt them appropriately. decrypted content is
delimited with ENC![]!<FORMAT> where FORMAT is one of STRING or
BLOCK depending on the format of the source encryption.