Fix the encrypt action
The action previously tried to use code for formatting the output (according
to the --output option) even when it was transforming a eyaml file. This made
no sense and resulting in a series of exceptions and problems.

Added a test that encrypts using the command that previously broke for me.