Collect configuration sources and log after parse
We do want to log where our configuration comes from as mentioned
in voxpupuli/hiera-eyaml#237 but that should take place on the `debug`
level and not the `info` level, since the behavior is documented in
the README and should only appear in verbose logs.

Additionally, that kind of logging should take place after all
the options have been parsed, since it matters what kind of verbosity
users want to have output.. and we do not know that until after
parsing. That results in a little bit of messy code where the sources
of the options are passed along with the options themselves, but also
means possible expansion in the future. Maybe.