(maint) Update Gemfile and testing for Puppet 6
Previously it was not possible to bundle install on Ruby 2.5 with Puppet 5 on
Windows due to FFI restrictions.  This commit updates the Gemfile to use the
appropriate default version of Puppet, based on the Ruby version

ActiveSupport (Which is used by github log generator) v6.x requires Ruby 2.5.
This commit adds a per-ruby version requirement for activesupport which will
allow it to be bundled correctly for Ruby 2.5 and 2.4.