Make public key file an RSA public key instead of a certificate
OpenSSL PKCS7 encrypt function needs a list of X509 certs to encrypt
and sign data. However, in the context of hiera-eyaml, the certificate
serves no purpose. For the user, it is easier to understand that eyaml
deals with a keypair. Internally, we wrap the public key in a X509 dummy
certificate to call PKCS7 encrypt but the user does not have to be exposed
to it.