Remove unneeded public certificate attributes
The following certificate attributes have no use in the
encryption or decryption process as it is implemented in hiera-eyaml:
- subject
- version
- digest type
- extensions

These attributes are required when executing with JRuby to sign the
certificate:
- begin date
- end date

The attributes that are always essential are the serial number and
the public key. When left unset, the serial number is generated
randomly or set to 0, which is fine in the context of hiera-eyaml.