Run CI in strict mode
By default, cucumber do not return an error when some step definitions
are missing: it only produce an informational message with code snippet
before exiting.

On CI, this is "lost" and when we break a test, it may seem the CI is
passing while in fact the test is just broken and not tested anymore.

When running in strict mode, cucumber will exit with an error in such
condition.