print prompts to stderr
Without this change, in the following command:

   eyaml encrypt -o string -p > file

...the "Enter password: " prompt is invisible to the user and is
redirected into the output file instead. If the user enters the password
blindly, then the password is encrypted and stored in the file. The file
still contains the "Enter password: ********\n" string, though, and that
output is printed upon:

   eyaml decrypt -f file

For consistency and general best practice, apply this change to all
user prompts. Include missing tests for createkeys prompt.

Extend cucumber to support testing with the necessary precision.